requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-cohere
langchain-google-genai
cohere
google-generativeai
PyMuPDF
python-docx
pytube
yt-dlp
beautifulsoup4
requests
pytesseract
pillow
unstructured
chromadb
tenacity
pydantic
tiktoken
ffmpeg-python
